requirements.txt
crawl4ai>=0.8.0
flask>=3.1.0
fastmcp>=3.1.0
python-dotenv>=1.0.0
httpx>=0.28.0
playwright>=1.50.0
